Use differential vector for features
The differential vector determines the separation between two teams
in a match. This way, two teams that are of the same quality are
treated the same, regardless of the level. For example, two top 25
teams will have a similar algorithm to two bottom 25 teams.
Additionally, teams that are of two different quality levels can be
clearly seen to theoritically expose the favorite.