Track if power conference (+1) member
Now including a feature which indicates whether or not a team is in
a power conference. Traditionally, the 'power conferences' have been
limited to the Big Ten, Big 12, SEC, ACC, and Pac-12, but due to the
recent showing of dominance and production of top-caliber teams, I am
including the Big East as a power conference team.