Find the projected NCAA tournament teams and seeds
Another feature should be added to forecast the NCAA tournament field
based on automatic and at-large bids. In order to populate the field, a
simulation of every conference tournament will be performed, and the
projected winner for each conference will receive an automatic bid. The
remaining teams will be selected in order based on the comparative power
rankings. Once all teams are selected, each team is assigned a seed
based on their position in the power rankings, with the first four teams
getting a 1 seed, the next four receiving a 2 seed, and so on.

Signed-Off-By: Robert Clark <robdclark@outlook.com>