Update name used for daily predicted winner
The team names listed in the home and away columns was different from
the winning team name which can be a bit confusing at first glance as it
might appear that a completely different team is the predicted winner in
some cases. Instead, the predicted winner should use the same name as
those in the home and away columns.

Signed-Off-By: Robert Clark <robdclark@outlook.com>