Fix issue with duplicate games in Monte Carlo
Remaining games in a schedule were being duplicated in the Monte Carlo
simulations when the algorithm attempted to determine if a game was
being played at the home or away location. Instead of comparing against
the string constant from sportsreference, the algorithm compared with a
string literal which was different from how sportsreference presented
the information. As such, every game was marked as a home game, so a
single matchup in the schedule was simulated twice - once where each
team was the home team.

Signed-Off-By: Robert Clark <robdclark@outlook.com>