Add script to build aggregated dataset
The process of reading all matches, converting them to pandas
DataFrames, removing duplicate items, and performing feature engineering
is time and compute-expensive. Instead, a simple script should be
created to automatically generate a singular dataset file which contains
all of the data necessary for prediction algorithms to be run. This
saves massive amounts of time and space, allowing workspaces to be
slimmer and predictions to be made quicker.

Signed-Off-By: Robert Clark <robdclark@outlook.com>