Calculate spreads instead of points
Predicting the final points tallies for the teams lead to drastic
under-fitting as the algorithm tended to pick scores in an average range
for typical games, and never lead to large point differences for heavily
outmatched opponents. Converting to spreads, a 1-dimensional vector,
will hopefully reduce under-fitting and help paint a more accurate
picture of the actual differences between teams.

Signed-Off-By: Robert Clark <robdclark@outlook.com>