Remove game pulling away from Docker
The Docker containers don't need to default to building a dataset as it
takes time and makes it appear commands are stalled. Instead, the
building of a dataset should happen outside of the container to make all
images as efficient as possible.

In addition, the sample dataset is no longer necessary as the repository
is intended to be used in the Docker container where a master dataset is
already included with the most complete number of features.

Signed-Off-By: Robert Clark <robdclark@outlook.com>