Update dataset based off existing dataset
The provided dataset in the Docker container should be used as a
baseline for creating a new dataset. This way, the dataset doesn't need
to be built from scratch every day, and a copy of all games in the
provided dataset don't need to be stored to a local filesystem.

Note: To pull the existing dataset from the most recent container, use
the following command:

docker run --rm -v $(pwd):/opt/mount --rm --entrypoint cp \
    roclark/clarktech-ncaab-predictor /app/dataset.pkl \
    /opt/mount/dataset.pkl

Signed-Off-By: Robert Clark <robdclark@outlook.com>