Only append new data to dataset
Instead of reading the final game on a team's schedule and checking if
it was played the previous day, the dataset builder should run through
the entire team's schedule and see if any games have not been added, and
add them if so. This would eliminate any race conditions where the data
isn't actually populated yet, and also avoid a scenario where the
builder would check the final game on a team's schedule prior to that
date actually occuring, and skip the most recent games.

Signed-Off-By: Robert Clark <robdclark@outlook.com>