Remove sorting of team's schedule
The attempt to sort the remainder of a team's schedule in the Monte
Carlo Simulation was causing errors given the sorting index is no longer
valid. The remaining schedules no longer need to be sorted as they
should already be in the correct order given how the data is presented.
Regardless, the order likely doesn't matter.

Signed-Off-By: Robert Clark <robdclark@outlook.com>