Create simulations directory to save simulations
Currently, while running the Monte Carlo Simulation, an error is thrown
at the end of the algorithm, indicating that the simulations/ directory
is not found, and it is therefore unable to save the result of the
simulation. To make matters easier, the simulation directory should be
created as part of the container creation steps to ensure it is always
present.

Signed-Off-By: Robert Clark <robdclark@outlook.com>